Reuse `effective_python_site_packages_path`
